api/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
dynamic = ["version"]
name = "makemkv-headless-api"
dependencies = [
    "bidict==0.23.1",
    "blinker==1.9.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.8",
    "deepmerge==2.0",
    "h11==0.16.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.2",
    "packaging==25.0",
    "python-engineio==4.12.2",
    "python-socketio==5.13.0",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "simple-websocket==1.1.0",
    "urllib3==2.2.3",
    "Werkzeug==3.1.3",
    "wsproto==1.2.0",
]

authors = [{ name = "enigma0z", email = "enigma.0za@gmail.com" }]
maintainers = [{ name = "enigma0z", email = "enigma.0za@gmail.com" }]

description = "Python wrapper-server built around makemkvcli and with built-in organization features for Jellyfin servers"
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = [
    "makemkv",
    "makemkv-headless",
    "makemkv-server",
    "jellyfin",
    "bd",
    "bdrom",
    "libbluray",
    "dvd",
    "rip",
    "video",
    "ffmpeg",
    "mkv",
]

[project.urls]
Homepage = "https://github.com/enigma0Z/makemkv-headless"
Documentation = "https://github.com/enigma0Z/makemkv-headless/blob/main/README.md"
Repository = "https://github.com/enigma0Z/makemkv-headless.git"
"Bug Tracker" = "https://github.com/enigma0Z/makemkv-headless/issues"
Changelog = "https://github.com/enigma0Z/makemkv-headless/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/makemkv_headless_api"]
